requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-dotenv==1.4.2
django-environ==0.11.2
django-unfold==0.43.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
MouseInfo==0.1.3
mysqlclient==2.2.6
outcome==1.3.0.post0
packaging==24.2
pi==0.1.2
pillow==11.1.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.32.3
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.8.2
wsproto==1.2.0
